requirements.txt
anyio==4.4.0
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
lxml==5.2.2
pyTelegramBotAPI==4.19.2
python-dotenv==1.0.1
python-telegram-bot==21.3
requests==2.31.0
sniffio==1.3.1
soupsieve==2.5
terminaltables==3.1.10
urllib3==2.2.1